* 'master' of jgit:AnalysisSuite:
  Fix bug in hlp() when roxygen comment starts from first line of file
  No point in showing startup message if not interactive
  Use path of startup script to set location of AnalysisSuite
  Process @ param ... as a parameter
  DOC/STY: document params
  Switch to error (rather than warning messages)
  Format params and other codes in hlp text
  Remove old code having found a way to extract roxygen comments
  Some left over work from an abortive attempt to use halp system for roxygen comments
  DOC: subset.neuronlist
  stop in subset.neuronlist when unprocessed args present
  STY: check for positive state of flag variable
  Fix bug in return nrrd version in is.nrrd
  Add ability to subset neuronlists by a function applied to each list item
  STY: blank line at end of file
  NrrdTestDataLength - a fast way to check that nrrd files are not truncated
  Teach Read3DDensityFromNrrd to read return header information only (ReadData=F)
  Fix bug in returning path of regular nrrd in NrrdDataFiles
  Add irtk.dofinvert function (for affine transforms)

+++ R/Code/NrrdFileFunctions.R
@@ -73,10 +73,11 @@ NrrdDataFiles<-function(nhdr,ReturnAbsPath=TRUE){
 		if(length(nhdr)>1) return(sapply(nhdr,NrrdDataFiles))
 		if(!is.nrrd(nhdr)) stop("This is not a nrrd file")
 		h=ReadNrrdHeader(nhdr)
-		if(is.null(h$datafile)) return(nhdr)
 	} else h=nhdr
-
-	if(length(h$datafile)>1){
+	if(is.null(h$datafile)){
+		# straight nrrd without detached header
+		dfs=attr(h,'path')
+	} else if(length(h$datafile)>1){
 		# list of files
 		dfs=h$datafile[-1]
 		firstlineparts=unlist(strsplit(h$datafile[1],"\\s+",perl=T))
@@ -92,6 +93,7 @@ NrrdDataFiles<-function(nhdr,ReturnAbsPath=TRUE){
 			seq(from=rangespec[1],to=rangespec[2],by=rangespec[3]))
 		if(length(firstlineparts)==5) attr(dfs,'subdim')=as.integer(firstlineparts[5])
 	} else dfs=h$datafile
+	
 	if(ReturnAbsPath){
 		# check if paths begin with /
 		relpaths=substring(dfs,1,1)!="/"
@@ -119,12 +121,15 @@ NrrdDataFiles<-function(nhdr,ReturnAbsPath=TRUE){
 	return(NULL)
 }
 
-Read3DDensityFromNrrd<-function(filename,Verbose=FALSE,AttachFullHeader=FALSE,
+Read3DDensityFromNrrd<-function(filename,Verbose=FALSE,ReadData=TRUE,AttachFullHeader=!ReadData,
 	ReadByteAsRaw=c("unsigned","all","none"),origin){
 	ReadByteAsRaw=match.arg(ReadByteAsRaw)
 	fc=file(filename,'rb')
 	h=ReadNrrdHeader(fc,CloseConnection=FALSE)
-		
+	# store the path because ReadNrrdHeader couldn't do it 
+	# TODO more elegant way of dealing with paths when connection sent to ReadNrrdHeader
+	attr(h,'path')=filename
+
 	# now read the data
 	dataTypes=data.frame(name=I(c("int8", "uint8", "int16", "uint16", "int32", "uint32", "int64", "uint64",
 	"float", "double", "block")),
@@ -153,30 +158,39 @@ Read3DDensityFromNrrd<-function(filename,Verbose=FALSE,AttachFullHeader=FALSE,
 	endian=ifelse(is.null(h$endian),.Platform$endian,h$endian)
 	if(Verbose) cat("dataLength =",dataLength,"dataType =",dataTypes$what[i],"size=",dataTypes$size[i],"\n")
 	enc=tolower(h$encoding)
-	if(enc=="raw"){
-		d=readBin(fc,what=dataTypes$what[i],n=dataLength,size=dataTypes$size[i],
-			signed=dataTypes$signed[i],endian=endian)
-		close(fc)
-	} else if(enc%in%c("gz","gzip")) {
-		# unfortunately gzcon seems to reset the connection 
-		# rather than starting to read from the current location
-		headerlength=seek(fc)
-		close(fc)
-		tf=tempfile()
-		system(paste('tail -c +',sep="",headerlength+1," ",filename," > ",tf))
-		gzf=gzfile(tf,'rb')
-		d=readBin(gzf,what=dataTypes$what[i],n=dataLength,size=dataTypes$size[i],
-			signed=dataTypes$signed[i],endian=endian)
-		close(gzf)
-		unlink(tf)
-	} else if(enc%in%c("ascii","txt","text")){
-		if(dataTypes$what[i]=='integer') whatVal=integer(0) else whatVal=double(0)
-		d=scan(fc,what=whatVal,nmax=dataLength,quiet=TRUE)
-		close(fc)
+	if(ReadData){
+		if(enc=="raw"){
+			d=readBin(fc,what=dataTypes$what[i],n=dataLength,size=dataTypes$size[i],
+				signed=dataTypes$signed[i],endian=endian)
+			close(fc)
+		} else if(enc%in%c("gz","gzip")) {
+			# unfortunately gzcon seems to reset the connection 
+			# rather than starting to read from the current location
+			headerlength=seek(fc)
+			close(fc)
+			tf=tempfile()
+			system(paste('tail -c +',sep="",headerlength+1," ",filename," > ",tf))
+			gzf=gzfile(tf,'rb')
+			d=readBin(gzf,what=dataTypes$what[i],n=dataLength,size=dataTypes$size[i],
+				signed=dataTypes$signed[i],endian=endian)
+			close(gzf)
+			unlink(tf)
+		} else if(enc%in%c("ascii","txt","text")){
+			if(dataTypes$what[i]=='integer') whatVal=integer(0) else whatVal=double(0)
+			d=scan(fc,what=whatVal,nmax=dataLength,quiet=TRUE)
+			close(fc)
+		} else {
+			stop("nrrd encoding ",enc," is not implemented")
+		}
+		dim(d)<-h$sizes
 	} else {
-		stop("nrrd encoding ",enc," is not implemented")
+		# don't read the data, we just wanted the (full) header information
+		if(dataTypes$what[i]=='integer') d=integer(0) else d=double(0)
+		attr(d,'datablock')=list(what=dataTypes$what[i],n=dataLength,size=dataTypes$size[i],
+			signed=dataTypes$signed[i],endian=endian)
+		attr(d,'datablock')$datastartpos=seek(fc)
+		close(fc)
 	}
-	dim(d)<-h$sizes
 	if(AttachFullHeader) attr(d,"header")=h
 	
 	if('space directions'%in%names(h)){
@@ -494,7 +508,7 @@ is.nrrd<-function(f,ReturnVersion=FALSE,TrustSuffix=FALSE){
 		return (FALSE)
 
 	if(ReturnVersion)
-		return(as.integer(magic[8]))
+		return(as.integer(magic[8])-0x30) # nb 0x30 is ASCII code for '0'
 
 	TRUE
 }
